Separated endpoint collating to separate function
